Added recognition of carousel flick transitions
Mainly for mobile, but also works for desktop; the
carousel recognises when images are moving at a high
velocity, which is usually an indicator of the user
wanting to flick to the next image.

Previously, when a small flick was done, it would
stay on the same image if you didn't flick far enough,
but now with small flicks it will move onto the next
image.